REFACTORING: removed folders per strct
