+++ src/decks/euchre24.rs
@@ -1,5 +1,5 @@
 use crate::types::rank::Rank;
-use crate::types::Ranked;
+use crate::types::traits::Ranked;
 
 #[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
 pub struct Euchre24 {}
